Thanh Son Push Notification

+++ rs-son/src/components/Navbar.vue
@@ -11,7 +11,10 @@
           </router-link>
         </li>
         <li v-if="tokenId != null && tokenId != ''">
-          <router-link to="/requests"> {{ $t("request") }} </router-link>
+          <router-link to="/requests" @click.prevent="deleteNotification">
+            {{ $t("request") }}
+            {{ notificaiton.length > 0 ? `+${notificaiton.length}` : "" }}
+          </router-link>
         </li>
         <li v-if="tokenId != null && tokenId != ''">
           <item-button @click="handleLogout()">{{ $t("logout") }}</item-button>
@@ -34,27 +37,71 @@
 
 <script>
 import { mapMutations, mapState } from "vuex";
+import firebase from "firebase/app";
 
 export default {
   data() {
     return {
       checkLogin: "",
       language: localStorage.getItem("lang") ? localStorage.getItem("lang") : "gb",
+      notificaiton: [],
     };
   },
   methods: {
     ...mapMutations("auth", ["SET_TOKEN_ID", "SET_LOCALE"]),
     handleLogout() {
-      localStorage.clear();
-      this.SET_TOKEN_ID("");
-      this.$store.commit("auth/SET_LOADING",false);
-      this.$router.push({ path: "/coaches" });
+      const db = firebase.firestore();
+      let CoachLogin = this.$store.state.auth.tokenId;
+      if (CoachLogin) {
+        localStorage.clear();
+        this.SET_TOKEN_ID("");
+        this.$store.commit("auth/SET_LOADING", false);
+        this.$router.push({ path: "/coaches" });
+        db.collection("user")
+          .get()
+          .then((querySnapshot) => {
+            querySnapshot.forEach((doc) => {
+              if (doc.data().idCoachLogin == CoachLogin.localId) {
+                db.collection("user").doc(`${doc.id}`).delete();
+              }
+            });
+          });
+      }
+    },
+    // GET NOTIFICATION FROM FIRESTORE
+    getNotification() {
+      const db = firebase.firestore();
+      let CoachLogin = this.$store.state.auth.tokenId;
+      db.collection("message")
+        .get()
+        .then((querySnapshot) => {
+          querySnapshot.forEach((doc) => {
+            if (CoachLogin) {
+              if (doc.data().content.idCoach == CoachLogin.localId) {
+                this.notificaiton.push(doc);
+              }
+            }
+          });
+        });
+    },
+    // DEL NOTIFICATION FROM FIRESTORE WHEN CLICK VIEW REQUEST
+    deleteNotification() {
+      const db = firebase.firestore();
+      this.notificaiton.forEach((item) => {
+        db.collection("message").doc(`${item.id}`).delete();
+        // SET TO 0
+        this.notificaiton = [];
+      });
+      this.$router.push("/requests");
     },
   },
   computed: {
     ...mapState(["auth"]),
     tokenId() {
       return this.auth.tokenId;
+    },
+    isNotification() {
+      return this.auth.isNotification;
     },
   },
   watch: {
@@ -63,6 +110,9 @@ export default {
       this.$i18n.locale = this.language;
       this.SET_LOCALE(this.language);
     },
+  },
+  mounted() {
+    this.getNotification();
   },
 };
 </script>
